Extract ConditionTranslator into its own file
